+++ public/js/user/dataTableInit.js
@@ -1,6 +1,10 @@
 $(document).ready(function() {
 	
 	$(".dataTable").dataTable({
+
+		fnPreDrawCallback: function () {
+			$(this).closest('.dataTables_wrapper').find( "input[type='search']").prop('spellcheck', false);
+		},
 		
 		stateSave: true,